start to split compiler.go
