+++ pkg/workflow/compiler.go
@@ -2170,561 +2170,14 @@ func (c *Compiler) buildAddReactionJob(data *WorkflowData, taskJobCreated bool, 
 	return job, nil
 }
 
-// buildCreateOutputIssueJob creates the create_issue job
-func (c *Compiler) buildCreateOutputIssueJob(data *WorkflowData, mainJobName string, taskJobCreated bool, frontmatter map[string]any) (*Job, error) {
-	if data.SafeOutputs == nil || data.SafeOutputs.CreateIssues == nil {
-		return nil, fmt.Errorf("safe-outputs.create-issue configuration is required")
-	}
 
-	var steps []string
 
-	// Add permission checks if no task job was created but permission checks are needed
-	if !taskJobCreated && c.needsPermissionChecks(data) {
-		// Add team member check step
-		steps = append(steps, "      - name: Check team membership for workflow\n")
-		steps = append(steps, "        id: check-team-member\n")
-		steps = append(steps, "        uses: actions/github-script@v8\n")
 
-		// Add environment variables for permission check
-		steps = append(steps, "        env:\n")
-		steps = append(steps, fmt.Sprintf("          GITHUB_AW_REQUIRED_ROLES: %s\n", strings.Join(data.Roles, ",")))
 
-		steps = append(steps, "        with:\n")
-		steps = append(steps, "          script: |\n")
 
-		// Generate the JavaScript code for the permission check
-		scriptContent := c.generatePermissionCheckScript(data.Roles)
-		scriptLines := strings.Split(scriptContent, "\n")
-		for _, line := range scriptLines {
-			if strings.TrimSpace(line) != "" {
-				steps = append(steps, fmt.Sprintf("            %s\n", line))
-			}
-		}
-	}
 
-	steps = append(steps, "      - name: Create Output Issue\n")
-	steps = append(steps, "        id: create_issue\n")
-	steps = append(steps, "        uses: actions/github-script@v8\n")
 
-	// Add environment variables
-	steps = append(steps, "        env:\n")
-	// Pass the agent output content from the main job
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_AGENT_OUTPUT: ${{ needs.%s.outputs.output }}\n", mainJobName))
-	if data.SafeOutputs.CreateIssues.TitlePrefix != "" {
-		steps = append(steps, fmt.Sprintf("          GITHUB_AW_ISSUE_TITLE_PREFIX: %q\n", data.SafeOutputs.CreateIssues.TitlePrefix))
-	}
-	if len(data.SafeOutputs.CreateIssues.Labels) > 0 {
-		labelsStr := strings.Join(data.SafeOutputs.CreateIssues.Labels, ",")
-		steps = append(steps, fmt.Sprintf("          GITHUB_AW_ISSUE_LABELS: %q\n", labelsStr))
-	}
 
-	// Pass the staged flag if it's set to true
-	if data.SafeOutputs.Staged != nil && *data.SafeOutputs.Staged {
-		steps = append(steps, "          GITHUB_AW_SAFE_OUTPUTS_STAGED: \"true\"\n")
-	}
-
-	// Add custom environment variables from safe-outputs.env
-	c.addCustomSafeOutputEnvVars(&steps, data)
-
-	steps = append(steps, "        with:\n")
-	// Add github-token if specified
-	var token string
-	if data.SafeOutputs.CreateIssues != nil {
-		token = data.SafeOutputs.CreateIssues.GitHubToken
-	}
-	c.addSafeOutputGitHubTokenForConfig(&steps, data, token)
-	steps = append(steps, "          script: |\n")
-
-	// Add each line of the script with proper indentation
-	formattedScript := FormatJavaScriptForYAML(createIssueScript)
-	steps = append(steps, formattedScript...)
-
-	// Create outputs for the job
-	outputs := map[string]string{
-		"issue_number": "${{ steps.create_issue.outputs.issue_number }}",
-		"issue_url":    "${{ steps.create_issue.outputs.issue_url }}",
-	}
-
-	// Determine the job condition for command workflows
-	var jobCondition string
-	if data.Command != "" {
-		// Build the command trigger condition
-		commandCondition := buildCommandOnlyCondition(data.Command)
-		commandConditionStr := commandCondition.Render()
-		jobCondition = commandConditionStr
-	} else {
-		jobCondition = "" // No conditional execution
-	}
-
-	// Set base permissions
-	permissions := "permissions:\n      contents: read\n      issues: write"
-
-	// Add actions: write permission if team member checks are present for command workflows
-	_, hasExplicitRoles := frontmatter["roles"]
-	requiresWorkflowCancellation := data.Command != "" ||
-		(!taskJobCreated && c.needsPermissionChecks(data) && hasExplicitRoles)
-
-	if requiresWorkflowCancellation {
-		permissions = "permissions:\n      actions: write  # Required for github.rest.actions.cancelWorkflowRun()\n      contents: read\n      issues: write"
-	}
-
-	job := &Job{
-		Name:           "create_issue",
-		If:             jobCondition,
-		RunsOn:         "runs-on: ubuntu-latest",
-		Permissions:    permissions,
-		TimeoutMinutes: 10, // 10-minute timeout as required
-		Steps:          steps,
-		Outputs:        outputs,
-		Needs:          []string{mainJobName}, // Depend on the main workflow job
-	}
-
-	return job, nil
-}
-
-// buildCreateOutputDiscussionJob creates the create_discussion job
-func (c *Compiler) buildCreateOutputDiscussionJob(data *WorkflowData, mainJobName string) (*Job, error) {
-	if data.SafeOutputs == nil || data.SafeOutputs.CreateDiscussions == nil {
-		return nil, fmt.Errorf("safe-outputs.create-discussion configuration is required")
-	}
-
-	var steps []string
-	steps = append(steps, "      - name: Create Output Discussion\n")
-	steps = append(steps, "        id: create_discussion\n")
-	steps = append(steps, "        uses: actions/github-script@v8\n")
-
-	// Add environment variables
-	steps = append(steps, "        env:\n")
-	// Pass the agent output content from the main job
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_AGENT_OUTPUT: ${{ needs.%s.outputs.output }}\n", mainJobName))
-	if data.SafeOutputs.CreateDiscussions.TitlePrefix != "" {
-		steps = append(steps, fmt.Sprintf("          GITHUB_AW_DISCUSSION_TITLE_PREFIX: %q\n", data.SafeOutputs.CreateDiscussions.TitlePrefix))
-	}
-	if data.SafeOutputs.CreateDiscussions.CategoryId != "" {
-		steps = append(steps, fmt.Sprintf("          GITHUB_AW_DISCUSSION_CATEGORY_ID: %q\n", data.SafeOutputs.CreateDiscussions.CategoryId))
-	}
-
-	// Pass the staged flag if it's set to true
-	if data.SafeOutputs.Staged != nil && *data.SafeOutputs.Staged {
-		steps = append(steps, "          GITHUB_AW_SAFE_OUTPUTS_STAGED: \"true\"\n")
-	}
-
-	// Add custom environment variables from safe-outputs.env
-	c.addCustomSafeOutputEnvVars(&steps, data)
-
-	steps = append(steps, "        with:\n")
-	// Add github-token if specified
-	var token string
-	if data.SafeOutputs.CreateDiscussions != nil {
-		token = data.SafeOutputs.CreateDiscussions.GitHubToken
-	}
-	c.addSafeOutputGitHubTokenForConfig(&steps, data, token)
-	steps = append(steps, "          script: |\n")
-
-	// Add each line of the script with proper indentation
-	formattedScript := FormatJavaScriptForYAML(createDiscussionScript)
-	steps = append(steps, formattedScript...)
-
-	outputs := map[string]string{
-		"discussion_number": "${{ steps.create_discussion.outputs.discussion_number }}",
-		"discussion_url":    "${{ steps.create_discussion.outputs.discussion_url }}",
-	}
-
-	// Determine the job condition based on command configuration
-	var jobCondition string
-	if data.Command != "" {
-		// Build the command trigger condition
-		commandCondition := buildCommandOnlyCondition(data.Command)
-		commandConditionStr := commandCondition.Render()
-		jobCondition = commandConditionStr
-	} else {
-		jobCondition = "" // No conditional execution
-	}
-
-	job := &Job{
-		Name:           "create_discussion",
-		If:             jobCondition,
-		RunsOn:         "runs-on: ubuntu-latest",
-		Permissions:    "permissions:\n      contents: read\n      discussions: write",
-		TimeoutMinutes: 10, // 10-minute timeout as required
-		Steps:          steps,
-		Outputs:        outputs,
-		Needs:          []string{mainJobName}, // Depend on the main workflow job
-	}
-
-	return job, nil
-}
-
-// buildCreateOutputAddCommentJob creates the create_issue_comment job
-func (c *Compiler) buildCreateOutputAddCommentJob(data *WorkflowData, mainJobName string) (*Job, error) {
-	if data.SafeOutputs == nil || data.SafeOutputs.AddComments == nil {
-		return nil, fmt.Errorf("safe-outputs.add-comment configuration is required")
-	}
-
-	var steps []string
-	steps = append(steps, "      - name: Add Issue Comment\n")
-	steps = append(steps, "        id: add_comment\n")
-	steps = append(steps, "        uses: actions/github-script@v8\n")
-
-	// Add environment variables
-	steps = append(steps, "        env:\n")
-	// Pass the agent output content from the main job
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_AGENT_OUTPUT: ${{ needs.%s.outputs.output }}\n", mainJobName))
-	// Pass the comment target configuration
-	if data.SafeOutputs.AddComments.Target != "" {
-		steps = append(steps, fmt.Sprintf("          GITHUB_AW_COMMENT_TARGET: %q\n", data.SafeOutputs.AddComments.Target))
-	}
-
-	// Add custom environment variables from safe-outputs.env
-	c.addCustomSafeOutputEnvVars(&steps, data)
-
-	steps = append(steps, "        with:\n")
-	// Add github-token if specified
-	var token string
-	if data.SafeOutputs.AddComments != nil {
-		token = data.SafeOutputs.AddComments.GitHubToken
-	}
-	c.addSafeOutputGitHubTokenForConfig(&steps, data, token)
-	steps = append(steps, "          script: |\n")
-
-	// Add each line of the script with proper indentation
-	formattedScript := FormatJavaScriptForYAML(createCommentScript)
-	steps = append(steps, formattedScript...)
-
-	// Create outputs for the job
-	outputs := map[string]string{
-		"comment_id":  "${{ steps.add_comment.outputs.comment_id }}",
-		"comment_url": "${{ steps.add_comment.outputs.comment_url }}",
-	}
-
-	// Determine the job condition based on target configuration
-	var baseCondition string
-	if data.SafeOutputs.AddComments.Target == "*" {
-		// Allow the job to run in any context when target is "*"
-		baseCondition = "always()" // This allows the job to run even without triggering issue/PR
-	} else {
-		// Default behavior: only run in issue or PR context
-		baseCondition = "github.event.issue.number || github.event.pull_request.number"
-	}
-
-	// If this is a command workflow, combine the command trigger condition with the base condition
-	var jobCondition string
-	if data.Command != "" {
-		// Build the command trigger condition
-		commandCondition := buildCommandOnlyCondition(data.Command)
-		commandConditionStr := commandCondition.Render()
-
-		// Combine command condition with base condition using AND
-		if baseCondition == "always()" {
-			// If base condition is always(), just use the command condition
-			jobCondition = commandConditionStr
-		} else {
-			// Combine both conditions with AND
-			jobCondition = fmt.Sprintf("(%s) && (%s)", commandConditionStr, baseCondition)
-		}
-	} else {
-		// No command trigger, just use the base condition
-		jobCondition = baseCondition
-	}
-
-	job := &Job{
-		Name:           "create_issue_comment",
-		If:             jobCondition,
-		RunsOn:         "runs-on: ubuntu-latest",
-		Permissions:    "permissions:\n      contents: read\n      issues: write\n      pull-requests: write",
-		TimeoutMinutes: 10, // 10-minute timeout as required
-		Steps:          steps,
-		Outputs:        outputs,
-		Needs:          []string{mainJobName}, // Depend on the main workflow job
-	}
-
-	return job, nil
-}
-
-// buildCreateOutputPullRequestReviewCommentJob creates the create_pr_review_comment job
-func (c *Compiler) buildCreateOutputPullRequestReviewCommentJob(data *WorkflowData, mainJobName string) (*Job, error) {
-	if data.SafeOutputs == nil || data.SafeOutputs.CreatePullRequestReviewComments == nil {
-		return nil, fmt.Errorf("safe-outputs.create-pull-request-review-comment configuration is required")
-	}
-
-	var steps []string
-	steps = append(steps, "      - name: Create PR Review Comment\n")
-	steps = append(steps, "        id: create_pr_review_comment\n")
-	steps = append(steps, "        uses: actions/github-script@v8\n")
-
-	// Add environment variables
-	steps = append(steps, "        env:\n")
-	// Pass the agent output content from the main job
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_AGENT_OUTPUT: ${{ needs.%s.outputs.output }}\n", mainJobName))
-	// Pass the side configuration
-	if data.SafeOutputs.CreatePullRequestReviewComments.Side != "" {
-		steps = append(steps, fmt.Sprintf("          GITHUB_AW_PR_REVIEW_COMMENT_SIDE: %q\n", data.SafeOutputs.CreatePullRequestReviewComments.Side))
-	}
-
-	// Add custom environment variables from safe-outputs.env
-	c.addCustomSafeOutputEnvVars(&steps, data)
-
-	steps = append(steps, "        with:\n")
-	// Add github-token if specified
-	var token string
-	if data.SafeOutputs.CreatePullRequestReviewComments != nil {
-		token = data.SafeOutputs.CreatePullRequestReviewComments.GitHubToken
-	}
-	c.addSafeOutputGitHubTokenForConfig(&steps, data, token)
-	steps = append(steps, "          script: |\n")
-
-	// Add each line of the script with proper indentation
-	formattedScript := FormatJavaScriptForYAML(createPRReviewCommentScript)
-	steps = append(steps, formattedScript...)
-
-	// Create outputs for the job
-	outputs := map[string]string{
-		"review_comment_id":  "${{ steps.create_pr_review_comment.outputs.review_comment_id }}",
-		"review_comment_url": "${{ steps.create_pr_review_comment.outputs.review_comment_url }}",
-	}
-
-	// We only run in pull request context, Note that in pull request comments only github.event.issue.pull_request is set.
-	baseCondition := "(github.event.issue.number && github.event.issue.pull_request) || github.event.pull_request"
-
-	// If this is a command workflow, combine the command trigger condition with the base condition
-	var jobCondition string
-	if data.Command != "" {
-		// Build the command trigger condition
-		commandCondition := buildCommandOnlyCondition(data.Command)
-		commandConditionStr := commandCondition.Render()
-
-		// Combine command condition with base condition using AND
-		jobCondition = fmt.Sprintf("(%s) && (%s)", commandConditionStr, baseCondition)
-	} else {
-		// No command trigger, just use the base condition
-		jobCondition = baseCondition
-	}
-
-	job := &Job{
-		Name:           "create_pr_review_comment",
-		If:             jobCondition,
-		RunsOn:         "runs-on: ubuntu-latest",
-		Permissions:    "permissions:\n      contents: read\n      pull-requests: write",
-		TimeoutMinutes: 10, // 10-minute timeout as required
-		Steps:          steps,
-		Outputs:        outputs,
-		Needs:          []string{mainJobName}, // Depend on the main workflow job
-	}
-
-	return job, nil
-}
-
-// buildCreateOutputCodeScanningAlertJob creates the create_code_scanning_alert job
-func (c *Compiler) buildCreateOutputCodeScanningAlertJob(data *WorkflowData, mainJobName string, workflowFilename string) (*Job, error) {
-	if data.SafeOutputs == nil || data.SafeOutputs.CreateCodeScanningAlerts == nil {
-		return nil, fmt.Errorf("safe-outputs.create-code-scanning-alert configuration is required")
-	}
-
-	var steps []string
-	steps = append(steps, "      - name: Create Code Scanning Alert\n")
-	steps = append(steps, "        id: create_code_scanning_alert\n")
-	steps = append(steps, "        uses: actions/github-script@v8\n")
-
-	// Add environment variables
-	steps = append(steps, "        env:\n")
-	// Pass the agent output content from the main job
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_AGENT_OUTPUT: ${{ needs.%s.outputs.output }}\n", mainJobName))
-	// Pass the max configuration
-	if data.SafeOutputs.CreateCodeScanningAlerts.Max > 0 {
-		steps = append(steps, fmt.Sprintf("          GITHUB_AW_SECURITY_REPORT_MAX: %d\n", data.SafeOutputs.CreateCodeScanningAlerts.Max))
-	}
-	// Pass the driver configuration, defaulting to frontmatter name
-	driverName := data.SafeOutputs.CreateCodeScanningAlerts.Driver
-	if driverName == "" {
-		if data.FrontmatterName != "" {
-			driverName = data.FrontmatterName
-		} else {
-			driverName = data.Name // fallback to H1 header name
-		}
-	}
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_SECURITY_REPORT_DRIVER: %s\n", driverName))
-	// Pass the workflow filename for rule ID prefix
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_WORKFLOW_FILENAME: %s\n", workflowFilename))
-
-	// Add custom environment variables from safe-outputs.env
-	c.addCustomSafeOutputEnvVars(&steps, data)
-
-	steps = append(steps, "        with:\n")
-	// Add github-token if specified
-	var token string
-	if data.SafeOutputs.CreateCodeScanningAlerts != nil {
-		token = data.SafeOutputs.CreateCodeScanningAlerts.GitHubToken
-	}
-	c.addSafeOutputGitHubTokenForConfig(&steps, data, token)
-	steps = append(steps, "          script: |\n")
-
-	// Add each line of the script with proper indentation
-	formattedScript := FormatJavaScriptForYAML(createCodeScanningAlertScript)
-	steps = append(steps, formattedScript...)
-
-	// Add step to upload SARIF artifact
-	steps = append(steps, "      - name: Upload SARIF artifact\n")
-	steps = append(steps, "        if: steps.create_code_scanning_alert.outputs.sarif_file\n")
-	steps = append(steps, "        uses: actions/upload-artifact@v4\n")
-	steps = append(steps, "        with:\n")
-	steps = append(steps, "          name: code-scanning-alert.sarif\n")
-	steps = append(steps, "          path: ${{ steps.create_code_scanning_alert.outputs.sarif_file }}\n")
-
-	// Add step to upload SARIF to GitHub Code Scanning
-	steps = append(steps, "      - name: Upload SARIF to GitHub Security\n")
-	steps = append(steps, "        if: steps.create_code_scanning_alert.outputs.sarif_file\n")
-	steps = append(steps, "        uses: github/codeql-action/upload-sarif@v3\n")
-	steps = append(steps, "        with:\n")
-	steps = append(steps, "          sarif_file: ${{ steps.create_code_scanning_alert.outputs.sarif_file }}\n")
-
-	// Create outputs for the job
-	outputs := map[string]string{
-		"sarif_file":        "${{ steps.create_code_scanning_alert.outputs.sarif_file }}",
-		"findings_count":    "${{ steps.create_code_scanning_alert.outputs.findings_count }}",
-		"artifact_uploaded": "${{ steps.create_code_scanning_alert.outputs.artifact_uploaded }}",
-		"codeql_uploaded":   "${{ steps.create_code_scanning_alert.outputs.codeql_uploaded }}",
-	}
-
-	// Build job condition - repository security advisories can run in any context unlike PR review comments
-	var jobCondition string
-	if data.Command != "" {
-		// Build the command trigger condition
-		commandCondition := buildCommandOnlyCondition(data.Command)
-		commandConditionStr := commandCondition.Render()
-		jobCondition = commandConditionStr
-	} else {
-		// No specific condition needed - repository security advisories can run anytime
-		jobCondition = ""
-	}
-
-	job := &Job{
-		Name:           "create_code_scanning_alert",
-		If:             jobCondition,
-		RunsOn:         "runs-on: ubuntu-latest",
-		Permissions:    "permissions:\n      contents: read\n      security-events: write\n      actions: read", // Need security-events:write for SARIF upload
-		TimeoutMinutes: 10,                                                                                      // 10-minute timeout
-		Steps:          steps,
-		Outputs:        outputs,
-		Needs:          []string{mainJobName}, // Depend on the main workflow job
-	}
-
-	return job, nil
-}
-
-// buildCreateOutputPullRequestJob creates the create_pull_request job
-func (c *Compiler) buildCreateOutputPullRequestJob(data *WorkflowData, mainJobName string) (*Job, error) {
-	if data.SafeOutputs == nil || data.SafeOutputs.CreatePullRequests == nil {
-		return nil, fmt.Errorf("safe-outputs.create-pull-request configuration is required")
-	}
-
-	var steps []string
-
-	// Step 1: Download patch artifact
-	steps = append(steps, "      - name: Download patch artifact\n")
-	steps = append(steps, "        continue-on-error: true\n")
-	steps = append(steps, "        uses: actions/download-artifact@v5\n")
-	steps = append(steps, "        with:\n")
-	steps = append(steps, "          name: aw.patch\n")
-	steps = append(steps, "          path: /tmp/\n")
-
-	// Step 2: Checkout repository
-	steps = append(steps, "      - name: Checkout repository\n")
-	steps = append(steps, "        uses: actions/checkout@v5\n")
-	steps = append(steps, "        with:\n")
-	steps = append(steps, "          fetch-depth: 0\n")
-
-	// Step 3: Configure Git credentials
-	steps = append(steps, c.generateGitConfigurationSteps()...)
-
-	// Step 4: Create pull request
-	steps = append(steps, "      - name: Create Pull Request\n")
-	steps = append(steps, "        id: create_pull_request\n")
-	steps = append(steps, "        uses: actions/github-script@v8\n")
-
-	// Add environment variables
-	steps = append(steps, "        env:\n")
-	// Pass the agent output content from the main job
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_AGENT_OUTPUT: ${{ needs.%s.outputs.output }}\n", mainJobName))
-	// Pass the workflow ID for branch naming
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_WORKFLOW_ID: %q\n", mainJobName))
-	// Pass the base branch from GitHub context
-	steps = append(steps, "          GITHUB_AW_BASE_BRANCH: ${{ github.ref_name }}\n")
-	if data.SafeOutputs.CreatePullRequests.TitlePrefix != "" {
-		steps = append(steps, fmt.Sprintf("          GITHUB_AW_PR_TITLE_PREFIX: %q\n", data.SafeOutputs.CreatePullRequests.TitlePrefix))
-	}
-	if len(data.SafeOutputs.CreatePullRequests.Labels) > 0 {
-		labelsStr := strings.Join(data.SafeOutputs.CreatePullRequests.Labels, ",")
-		steps = append(steps, fmt.Sprintf("          GITHUB_AW_PR_LABELS: %q\n", labelsStr))
-	}
-	// Pass draft setting - default to true for backwards compatibility
-	draftValue := true // Default value
-	if data.SafeOutputs.CreatePullRequests.Draft != nil {
-		draftValue = *data.SafeOutputs.CreatePullRequests.Draft
-	}
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_PR_DRAFT: %q\n", fmt.Sprintf("%t", draftValue)))
-
-	// Pass the if-no-changes configuration
-	ifNoChanges := data.SafeOutputs.CreatePullRequests.IfNoChanges
-	if ifNoChanges == "" {
-		ifNoChanges = "warn" // Default value
-	}
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_PR_IF_NO_CHANGES: %q\n", ifNoChanges))
-
-	// Pass the maximum patch size configuration
-	maxPatchSize := 1024 // Default value
-	if data.SafeOutputs != nil && data.SafeOutputs.MaximumPatchSize > 0 {
-		maxPatchSize = data.SafeOutputs.MaximumPatchSize
-	}
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_MAX_PATCH_SIZE: %d\n", maxPatchSize))
-
-	// Pass the staged flag if it's set to true
-	if data.SafeOutputs.Staged != nil && *data.SafeOutputs.Staged {
-		steps = append(steps, "          GITHUB_AW_SAFE_OUTPUTS_STAGED: \"true\"\n")
-	}
-
-	// Add custom environment variables from safe-outputs.env
-	c.addCustomSafeOutputEnvVars(&steps, data)
-
-	steps = append(steps, "        with:\n")
-	// Add github-token if specified
-	c.addSafeOutputGitHubTokenForConfig(&steps, data, data.SafeOutputs.CreatePullRequests.GitHubToken)
-	steps = append(steps, "          script: |\n")
-
-	// Add each line of the script with proper indentation
-	formattedScript := FormatJavaScriptForYAML(createPullRequestScript)
-	steps = append(steps, formattedScript...)
-
-	// Create outputs for the job
-	outputs := map[string]string{
-		"pull_request_number": "${{ steps.create_pull_request.outputs.pull_request_number }}",
-		"pull_request_url":    "${{ steps.create_pull_request.outputs.pull_request_url }}",
-		"branch_name":         "${{ steps.create_pull_request.outputs.branch_name }}",
-	}
-
-	// Determine the job condition for command workflows
-	var jobCondition string
-	if data.Command != "" {
-		// Build the command trigger condition
-		commandCondition := buildCommandOnlyCondition(data.Command)
-		commandConditionStr := commandCondition.Render()
-		jobCondition = commandConditionStr
-	} else {
-		jobCondition = "" // No conditional execution
-	}
-
-	job := &Job{
-		Name:           "create_pull_request",
-		If:             jobCondition,
-		RunsOn:         "runs-on: ubuntu-latest",
-		Permissions:    "permissions:\n      contents: write\n      issues: write\n      pull-requests: write",
-		TimeoutMinutes: 10, // 10-minute timeout as required
-		Steps:          steps,
-		Outputs:        outputs,
-		Needs:          []string{mainJobName}, // Depend on the main workflow job
-	}
-
-	return job, nil
-}
 
 // buildMainJob creates the main workflow job
 func (c *Compiler) buildMainJob(data *WorkflowData, jobName string, taskJobCreated bool, frontmatter map[string]any) (*Job, error) {
